requirements.txt
pyserial
flask
flask-restful
